requirements.txt
adjustText==1.3.0
arxiv==2.1.3
bcrypt==4.2.0
deepl==1.21.1
fake_useragent==2.0.3
Flask==3.1.0
Flask_Cors==5.0.0
Flask_SocketIO==5.3.7
flask_talisman==1.1.0
fuzzywuzzy==0.18.0
gymnasium==1.0.0
habanero==2.2.0
hypergraphx==1.7.5
imbalanced_learn==0.12.2
imblearn==0.0
infomap==2.8.0
keyboard==0.13.5
langchain_community==0.3.20
lightgbm==4.6.0
matplotlib==3.8.4
mysql-connector-python==8.4.0
mysql_connector_repackaged==0.3.1
networkx==3.3
numpy==1.26.2
openai==1.70.0
opencv_python==4.9.0.80
optuna==4.0.0
pandas==2.0.3
Pillow==10.3.0
PyWavelets==1.6.0
plotly==5.22.0
PuLP==2.9.0
pyarrow==17.0.0
PyJWT==2.8.0
pypdf==5.2.0
python-dotenv==1.1.0
python-Levenshtein==0.26.1
Requests==2.32.3
scikit_learn==1.4.2
scipy==1.15.2
selenium==4.30.0
stable_baselines3==2.4.1
textblob==0.19.0
tiktoken==0.8.0
torch==2.3.0
torch_geometric==2.6.1
tqdm==4.67.1
umap_learn==0.5.7
urllib3==2.2.0
waitress==3.0.2
zhipuai==2.1.5.20250106
gunicorn==23.0.0
